EagleDev2025_HateSpeechBot/backend/requirements.txt
# Backend
flask
pymongo

# Gemini (Google Generative AI)
google-generativeai

# Procesamiento de audio
openai-whisper
ffmpeg-python

# IA / ML
tensorflow

# Utilidades
pandas
python-dotenv